No mhash extension past php5.3
From: http://php.net/manual/en/mhash.installation.php

As of PHP 5.3.0, the mhash extension is emulated thru the Hash extension. This makes the mhash installation directory have no effect, and requires the hash extension enabled to make the mhash support available.

mhash is not a valid extension for php 5.6, there is only hash. This is misleading.